+++ R/internal.R
@@ -172,7 +172,7 @@ unlist_variables <- function(var_names){
 #' @description
 #' Checks for valid character vectors.
 #'
-#' @param vector_to_check A vector to check..
+#' @param vector_to_check A vector to check.
 #'
 #' @return
 #' Returns TRUE or FALSE depending on the result.
@@ -180,4 +180,27 @@ unlist_variables <- function(var_names){
 #' @noRd
 is_valid_vector <- function(vector_to_check){
     length(vector_to_check) > 1
+}
+
+
+#' Catch Specific Message
+#'
+#' @description
+#' Catches a specific message and excludes it from the output while printing all
+#' other messages as normal.
+#'
+#' @param call_function A function which outputs messages.
+#' @param pattern The message pattern for which to look out for.
+#'
+#' @return
+#' Returns the call_functions return value.
+#'
+#' @noRd
+suppress_specific_message <- function(call_function, pattern) {
+    withCallingHandlers(call_function,
+                        message = function(single_message) {
+                            if (grepl(pattern, conditionMessage(single_message))){
+                                invokeRestart("muffleMessage")
+                            }
+                        })
 }