+++ tests/testthat/test-import_export.R
@@ -1,4 +1,4 @@
-test_that("Example files load", {
+test_that("Example csv and xlsx files load", {
     csv_file  <- system.file("extdata", "qol_example_data_csv.csv",  package = "qol")
     xlsx_file <- system.file("extdata", "qol_example_data_xlsx.xlsx", package = "qol")